Issue #839 Move widely used test helper file
