+++ src/locale/__tests__/locale.test.ts
@@ -1,6 +1,6 @@
 // tslint:disable:no-expression-statement
 import { LocaleInfoManager } from '..';
-import { aString } from '../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../helpers/random_test_values';
 import { LocaleInfoBuilder } from '../../stores/__tests__/helpers/locale_helpers';
 import { LocaleInfo, CatalogsMap } from '../types';
