+++ src/selectors/__tests__/get_tasks_by_explore_taxonomy_term.test.ts
@@ -1,6 +1,6 @@
 // tslint:disable:no-expression-statement no-let
 import { ExploreTaxonomyId, TaxonomyTermReference } from '../../stores/taxonomies';
-import { aString } from '../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../helpers/random_test_values';
 import { TopicBuilder } from '../../stores/__tests__/helpers/topics_helpers';
 import { TopicMap } from '../../stores/topics';
 import { findItemsByExploreTaxonomyTerm } from '../taxonomies/find_items_by_explore_taxonomy_term';
