+++ src/selectors/__tests__/helpers/task_list_item_builder.ts
@@ -1,7 +1,7 @@
 // tslint:disable:no-class no-this readonly-keyword no-expression-statement
 
 import { ExploreSection } from '../../explore/types';
-import { aString, aBoolean } from '../../../application/__tests__/helpers/random_test_values';
+import { aString, aBoolean } from '../../../helpers/random_test_values';
 import { TopicListItem } from '../../topics/topic_list_item';
 
 export class TaskListItemBuilder {
