+++ src/selectors/__tests__/locale.test.ts
@@ -2,7 +2,7 @@
 import { LocaleStoreBuilder, LocalizedTextBuilder, aLocale } from '../../stores/__tests__/helpers/locale_helpers';
 import { getLocalizedText, getLocalizedTextOrDefault } from '../locale/get_localized_text';
 import * as locale from '../../stores/locale';
-import { aString } from '../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../helpers/random_test_values';
 import { toSelectorLocale } from '../locale/to_selector_locale';
 
 describe('locale selectors fetch', () => {
