+++ src/selectors/__tests__/select_icon_from_explore_taxonomy.test.ts
@@ -3,7 +3,7 @@
 
 import { selectIconFromExploreTaxonomy } from '../explore/select_icon_from_explore_taxonomy';
 import { ExploreTaxonomyId, TaxonomyTermReference, Taxonomy } from '../../stores/taxonomies';
-import { aString } from '../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../helpers/random_test_values';
 
 const taxonomyRefsWithExploreTerm = (term: string): ReadonlyArray<TaxonomyTermReference> => ([
     {
