+++ src/stores/__tests__/fonts.test.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-expression-statement
-import { aString } from '../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../helpers/random_test_values';
 import * as fonts from '../fonts';
 import * as constants from '../../application/constants';
