+++ src/stores/__tests__/helpers/explore_section_builder.ts
@@ -1,7 +1,7 @@
 // tslint:disable:no-class readonly-keyword no-this no-expression-statement
 import { Id, ExploreSection } from '../../explore';
 import { TaxonomyTermReference } from '../../taxonomies';
-import { aString } from '../../../application/__tests__/helpers/random_test_values';
+import { aString } from '../../../helpers/random_test_values';
 
 export class ExploreSectionBuilder {
     id: Id = aString();
