+++ src/stores/__tests__/helpers/services_helpers.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-class no-this readonly-keyword no-expression-statement
-import { aString, aNumber } from '../../../application/__tests__/helpers/random_test_values';
+import { aString, aNumber } from '../../../helpers/random_test_values';
 import { Id } from '../../services';
 import { Errors } from '../../../validation/errors/types';
 import {
