+++ src/validation/services/__tests__/helpers/services_schema_helpers.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-class readonly-keyword no-expression-statement no-this no-let no-any
-import { aString, anInteger, aNumber } from '../../../../application/__tests__/helpers/random_test_values';
+import { aString, anInteger, aNumber } from '../../../../helpers/random_test_values';
 
 // The following "JSON" types exist purely to allow us to test our schemas with invalid data.
 // The optional and any properties on these types ensure we can create invalid objects.
